강의실

+++ hello-spring/src/main/java/com/ktdsuniversity/edu/board/service/BoardServiceImpl.java
@@ -14,10 +14,11 @@ import org.springframework.web.multipart.MultipartFile;
 import com.ktdsuniversity.edu.board.dao.BoardDao;
 import com.ktdsuniversity.edu.board.enums.ReadType;
 import com.ktdsuniversity.edu.board.vo.BoardVO;
-import com.ktdsuniversity.edu.board.vo.SearchResultVO;
 import com.ktdsuniversity.edu.board.vo.request.SearchListVO;
 import com.ktdsuniversity.edu.board.vo.request.UpdateVO;
 import com.ktdsuniversity.edu.board.vo.request.WriteVO;
+import com.ktdsuniversity.edu.board.vo.response.SearchResultVO;
+import com.ktdsuniversity.edu.common.utils.AuthUtils;
 import com.ktdsuniversity.edu.exceptions.HelloSpringException;
 import com.ktdsuniversity.edu.files.dao.FilesDao;
 import com.ktdsuniversity.edu.files.helpers.MultipartFileHandler;
@@ -49,7 +50,7 @@ public class BoardServiceImpl implements BoardService {
 		int count = this.boardDao.selectBoardCount(searchListVO);
 		result.setCount(count);
 
-		
+		// 몇 개의 페이지가 필요한지 계산.
 		searchListVO.computePagination(count);
 		
 		if (count == 0) {
@@ -90,6 +91,9 @@ public class BoardServiceImpl implements BoardService {
 	@Override
 	public BoardVO findBoardByArticleId(String articleId, ReadType readType) {
 		if (readType == ReadType.VIEW) { 
+			
+
+
 			// 1. 조회수 증가.
 			int updateCount = this.boardDao.updateViewCntIncreaseById(articleId);
 			logger.debug("조회수가 증가된 게시글의 수: {}", updateCount);
@@ -102,6 +106,15 @@ public class BoardServiceImpl implements BoardService {
 		
 		// 2. 게시글 조회.
 		BoardVO board = this.boardDao.selectBoardById(articleId);
+		if(readType == ReadType.UPDATE) {
+			
+			String loginUserEmail = AuthUtils.getUsername();
+			boolean isAdminAccount = AuthUtils.hasAnyRole("RL-20260418-000001", "RL-20260418-000002");
+
+			if (!isAdminAccount && !loginUserEmail.equals(board.getEmail())) {
+				throw new HelloSpringException("잘못된 접근입니다.", "errors/403");
+			}
+		}
 		
 		// 조회한 게시글을 반환.
 		return board;
@@ -110,6 +123,14 @@ public class BoardServiceImpl implements BoardService {
 	@Transactional
 	@Override
 	public boolean deleteBoardByArticleId(String id) {
+		
+		BoardVO board = this.boardDao.selectBoardById(id);
+		String loginUserEmail = AuthUtils.getUsername();
+		boolean isAdminAccount = AuthUtils.hasAnyRole("RL-20260418-000001", "RL-20260418-000002");
+		if (!isAdminAccount && !loginUserEmail.equals(board.getEmail())) {
+			throw new HelloSpringException("잘못된 접근입니다.", "errors/403");
+		}
+		
 		int deleteCount = this.boardDao.deleteBoardById(id);
 		
 		// 삭제하려는 게시글에 첨부된 파일 목록을 가져온다.
@@ -132,6 +153,19 @@ public class BoardServiceImpl implements BoardService {
 	@Transactional
 	@Override
 	public boolean updateBoardByArticleId(UpdateVO updateVO) {
+		
+		BoardVO board = this.boardDao.selectBoardById(updateVO.getId());
+		// 게시글을 불러오고
+		
+		// 권한 검사 한 이후에 경우에 따라 예외를 던져준다.
+		String loginUserEmail = AuthUtils.getUsername();
+		boolean isAdminAccount = AuthUtils.hasAnyRole("RL-20260418-000001", "RL-20260418-000002");
+		if (!isAdminAccount && !loginUserEmail.equals(board.getEmail())) {
+			throw new HelloSpringException("잘못된 접근입니다.", "errors/403");
+		}
+		
+		
+		
 		// 선택한 파일들만 삭제.
 		if ( updateVO.getDeleteFileNum() != null && 
 				updateVO.getDeleteFileNum().size() > 0) {
@@ -168,6 +202,15 @@ public class BoardServiceImpl implements BoardService {
 		return updateCount == 1;
 	}
 
+	@Override
+	public boolean deleteAllBoardArticle(String id) {
 
+		boolean isAdminAccount = AuthUtils.hasAnyRole("RL-20260418-000001", "RL-20260418-000002");
+		if(!isAdminAccount) {
+			throw new HelloSpringException("잘못된 접근입니다.", "errors/403");	
+		}
+		boolean deleteResult = this.boardDao.deleteBoardAll();
+		return deleteResult;
+	}
 
 }
